un-conflate different views
