+++ packages/web/app/routes/AppointmentRoutes.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import { Redirect, Route, Switch } from 'react-router-dom';
 
-import { AppointmentsCalendar, AppointmentsCalendarOld } from '../views/scheduling';
+import { AppointmentsCalendar, LocationBookingsView } from '../views/scheduling';
 
 export const AppointmentRoutes = React.memo(({ match }) => (
   <Switch>
     <Route path={`${match.path}/calendar`} component={AppointmentsCalendar} />
-    <Route path={`${match.path}/old-calendar`} component={AppointmentsCalendarOld} />
+    <Route path={`${match.path}/locations`} component={LocationBookingsView} />
     <Redirect to={`${match.path}/calendar`} />
   </Switch>
 ));
