+++ packages/web/app/views/scheduling/AppointmentsCalendar.jsx
@@ -1,254 +1,259 @@
-import Chance from 'chance';
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
+import { add, endOfDay, format, startOfDay } from 'date-fns';
+import { ButtonGroup, IconButton, Typography } from '@material-ui/core';
+import ArrowBackIcon from '@material-ui/icons/KeyboardArrowLeft';
+import ArrowForwardIcon from '@material-ui/icons/KeyboardArrowRight';
 
+import { toDateTimeString } from '@tamanu/shared/utils/dateTime';
+import { PageContainer, TOP_BAR_HEIGHT, TopBar as TopBarBase } from '../../components';
+import { TwoColumnDisplay } from '../../components/TwoColumnDisplay';
+import { DailySchedule } from '../../components/Appointments/DailySchedule';
+import { NewAppointmentButton } from '../../components/Appointments/NewAppointmentButton';
+import { Button } from '../../components/Button';
+import { AutocompleteInput, TranslatedMultiSelectField } from '../../components/Field';
+import { Suggester } from '../../utils/suggester';
 import { Colors } from '../../constants';
-import { useApi } from '../../api';
-import { PageContainer, TOP_BAR_HEIGHT, TopBar, TranslatedText } from '../../components';
-import { CalendarCell } from './CalendarCell';
+import { useApi, useSuggester } from '../../api';
+import { TranslatedText } from '../../components/Translation/TranslatedText';
+import { useAuth } from '../../contexts/Auth';
+import { ErrorMessage } from '../../components/ErrorMessage';
+import { APPOINTMENT_TYPE_LABELS } from '@tamanu/constants';
 
-// BEGIN PLACEHOLDERS
-
-const chance = new Chance();
-
-const locations = Array.from({ length: 8 }, () => ({
-  id: chance.guid(),
-  code: chance.syllable(),
-  name: chance.sentence({ words: 3 }).slice(0, -1),
-  locationGroupId: chance.guid(),
-}));
-
-const Placeholder = styled.div`
-  background-color: oklch(0% 0 0 / 3%);
-  block-size: 100%;
-  border: 1px solid oklch(0% 0 0 / 15%);
-  color: oklch(0% 0 0 / 55%);
-  display: grid;
-  font-size: 1rem;
-  inline-size: 100%;
-  padding: 0.5rem;
-  place-items: center;
-  text-align: center;
+const LeftContainer = styled.div`
+  min-height: 100%;
+  border-right: 1px solid ${Colors.outline};
 `;
 
-// END PLACEHOLDERS
-
-const Wrapper = styled(PageContainer)`
-  display: grid;
-  grid-template-rows: min-content auto;
+const RightContainer = styled.div`
+  position: relative;
 `;
 
-const Filters = styled('search')`
+const TopBar = styled(TopBarBase)`
+  position: sticky;
+  top: 0;
+  height: ${TOP_BAR_HEIGHT}px;
+`;
+
+const DateHeader = styled.div`
   display: flex;
-  gap: 1rem;
+  align-items: center;
 `;
 
-const GridCarousel = styled.div``;
-
-const CalendarGrid = styled.div`
-  --location-count: 10;
-  --week-count: 6; // A month can span 4–6 distinct ISO weeks
-  --day-count: calc(var(--week-count) * 7);
-
-  --header-col-width: 8rem;
-  --col-width: calc((100cqi - var(--header-col-width)) / 7.5);
-
-  background-color: ${Colors.white};
-  border: max(0.0625rem, 1px) solid ${Colors.outline};
-  display: grid;
-  grid-auto-flow: column;
-  grid-template-columns: 8rem repeat(calc(var(--day-count), var(--col-width)));
-  grid-template-rows: repeat(var(--location-count), minmax(2.25rem, max-content));
-  margin: 1rem;
-  scroll-snap-type: inline mandatory;
+const DateDisplay = styled.span`
+  font-size: 1.2em;
+  margin: 0 12px;
+  min-width: 250px;
+  display: flex;
+  justify-content: center;
+  color: ${Colors.darkText};
 `;
 
-const TableCarousel = styled.table`
-  --location-count: 10;
-  --week-count: 6; // A month can span 4–6 distinct ISO weeks
+const CalendarContainer = styled.div`
+  overflow: auto;
+  height: calc(100vh - ${TOP_BAR_HEIGHT}px - 1px);
+  width: 100%;
+  position: absolute;
+`;
 
-  --header-col-width: 8rem;
-  --col-width: calc((100cqi - var(--header-col-width)) / 7.5);
+const Section = styled.div`
+  padding: 20px;
+  border-bottom: 1px solid ${Colors.outline};
+  display: flex;
+  flex-direction: column;
+`;
 
-  border-collapse: collapse;
-  background-color: ${Colors.white};
-  border: max(0.0625rem, 1px) solid ${Colors.outline};
-  margin: 1rem;
+const SectionTitle = styled(Typography)`
+  margin-bottom: 1rem;
+`;
+
+const FilterSwitch = styled(ButtonGroup)`
+  margin-top: 0.5rem;
+`;
+
+const NavigationButton = styled(IconButton)`
+  color: ${Colors.darkText};
+  padding: 4px;
+`;
+
+const TodayButton = styled(Button)`
+  margin-right: 12px;
+  & span {
+    text-decoration: underline;
+  }
 `;
 
 export const AppointmentsCalendar = () => {
-  // const api = useApi();
-  const appointments = [
-    {
-      id: '8c417eb4-d8f2-441f-91ed-2e5de7e89ee8',
-      startTime: '2024-09-12 11:48:00',
-      endTime: null,
-      type: 'Standard',
-      status: 'Confirmed',
-      updatedAtSyncTick: '35538',
-      createdAt: '2024-09-11T23:48:22.958Z',
-      updatedAt: '2024-09-11T23:48:22.958Z',
-      deletedAt: null,
-      patientId: '41f97737-be54-4e43-ba04-78b39656c30b',
-      clinicianId: '11111111-1111-1111-1111-111111111111',
-      locationGroupId: 'locationgroup-Theatre1-tamanu',
-      locationId: null,
-      patient: {
-        id: '41f97737-be54-4e43-ba04-78b39656c30b',
-        displayId: 'RQLN820387',
-        firstName: 'Adeline',
-        middleName: 'Ora',
-        lastName: 'Giannoni',
-        culturalName: 'Hattie',
-        dateOfBirth: '1998-12-20',
-        dateOfDeath: null,
-        sex: 'female',
-        email: null,
-        visibilityStatus: 'current',
-        updatedAtSyncTick: '-999',
-        createdAt: '2024-06-24T00:11:13.082Z',
-        updatedAt: '2024-06-24T00:11:13.082Z',
-        deletedAt: null,
-        villageId: 'village-Matacaucau',
-        mergedIntoId: null,
-        village: {
-          id: 'village-Matacaucau',
-          code: 'Matacaucau',
-          type: 'village',
-          name: 'Matacaucau',
-          visibilityStatus: 'current',
-          updatedAtSyncTick: '-999',
-          createdAt: '2024-06-24T00:11:01.546Z',
-          updatedAt: '2024-06-24T00:11:01.546Z',
-          deletedAt: null,
-        },
-      },
-      clinician: {
-        id: '11111111-1111-1111-1111-111111111111',
-        displayId: null,
-        email: 'jasper@bes.au',
-        displayName: 'Jasper Lai',
-        role: 'admin',
-        phoneNumber: null,
-        visibilityStatus: 'current',
-        updatedAtSyncTick: '35406',
-        createdAt: '2024-08-10T00:11:17.557Z',
-        updatedAt: '2024-09-11T23:03:33.526Z',
-        deletedAt: null,
-      },
-      locationGroup: {
-        id: 'locationgroup-Theatre1-tamanu',
-        code: 'Theatre1',
-        name: 'Operating Theatre',
-        visibilityStatus: 'current',
-        updatedAtSyncTick: '6528',
-        createdAt: '2024-06-24T00:11:12.974Z',
-        updatedAt: '2024-07-08T23:17:02.816Z',
-        deletedAt: null,
-        facilityId: 'facility-a',
-      },
-    },
-    {
-      id: '009f844a-20e1-47e0-aa82-30ab92017255',
-      startTime: '2024-09-12 11:49:00',
-      endTime: null,
-      type: 'Specialist',
-      status: 'Confirmed',
-      updatedAtSyncTick: '35542',
-      createdAt: '2024-09-11T23:49:49.375Z',
-      updatedAt: '2024-09-11T23:49:49.375Z',
-      deletedAt: null,
-      patientId: 'f8e332bb-6366-44f2-8254-1f10e4bd3bf7',
-      clinicianId: '11111111-1111-1111-1111-111111111111',
-      locationGroupId: 'locationgroup-Theatre1-tamanu',
-      locationId: null,
-      patient: {
-        id: 'f8e332bb-6366-44f2-8254-1f10e4bd3bf7',
-        displayId: 'TJOR564025',
-        firstName: 'Andre',
-        middleName: 'Elijah',
-        lastName: 'Tinti',
-        culturalName: 'Samuel',
-        dateOfBirth: '1941-09-04',
-        dateOfDeath: null,
-        sex: 'male',
-        email: null,
-        visibilityStatus: 'current',
-        updatedAtSyncTick: '-999',
-        createdAt: '2024-06-24T00:11:13.082Z',
-        updatedAt: '2024-06-24T00:11:13.082Z',
-        deletedAt: null,
-        villageId: 'village-Nasau',
-        mergedIntoId: null,
-        village: {
-          id: 'village-Nasau',
-          code: 'Nasau',
-          type: 'village',
-          name: 'Nasau',
-          visibilityStatus: 'current',
-          updatedAtSyncTick: '-999',
-          createdAt: '2024-06-24T00:11:06.577Z',
-          updatedAt: '2024-06-24T00:11:06.577Z',
-          deletedAt: null,
-        },
-      },
-      clinician: {
-        id: '11111111-1111-1111-1111-111111111111',
-        displayId: null,
-        email: 'jasper@bes.au',
-        displayName: 'Jasper Lai',
-        role: 'admin',
-        phoneNumber: null,
-        visibilityStatus: 'current',
-        updatedAtSyncTick: '35406',
-        createdAt: '2024-08-10T00:11:17.557Z',
-        updatedAt: '2024-09-11T23:03:33.526Z',
-        deletedAt: null,
-      },
-      locationGroup: {
-        id: 'locationgroup-Theatre1-tamanu',
-        code: 'Theatre1',
-        name: 'Operating Theatre',
-        visibilityStatus: 'current',
-        updatedAtSyncTick: '6528',
-        createdAt: '2024-06-24T00:11:12.974Z',
-        updatedAt: '2024-07-08T23:17:02.816Z',
-        deletedAt: null,
-        facilityId: 'facility-a',
-      },
-    },
-  ];
+  const api = useApi();
+  const locationGroupSuggester = useSuggester('facilityLocationGroup');
+  const { ability } = useAuth();
 
-  console.log('locations', locations);
+  const [date, setDate] = useState(new Date());
+  const [filterValue, setFilterValue] = useState('');
+  const [appointmentType, setAppointmentType] = useState([]);
+  const [appointments, setAppointments] = useState([]);
+  const [refreshCount, setRefreshCount] = useState(0);
+  const [activeFilter, setActiveFilter] = useState('locationGroup');
+
+  const hasPermission = ability.can('list', 'Appointment');
+
+  const updateCalendar = () => {
+    setRefreshCount(refreshCount + 1);
+  };
+  const updateFilterValue = e => setFilterValue(e.target.value || '');
+
+  const filters = {
+    locationGroup: {
+      label: <TranslatedText stringId="general.area.label" fallback="Area" />,
+      component: (
+        <AutocompleteInput
+          value={filterValue}
+          onChange={updateFilterValue}
+          suggester={locationGroupSuggester}
+        />
+      ),
+    },
+    clinician: {
+      label: (
+        <TranslatedText
+          stringId="general.localisedField.clinician.label.plural"
+          fallback="Clinicians"
+        />
+      ),
+      component: (
+        <AutocompleteInput
+          value={filterValue}
+          onChange={updateFilterValue}
+          suggester={new Suggester(api, 'practitioner')}
+        />
+      ),
+    },
+  };
+
+  useEffect(() => {
+    (async () => {
+      const { data } = await api.get('appointments', {
+        after: toDateTimeString(startOfDay(date)),
+        before: toDateTimeString(endOfDay(date)),
+        all: true,
+      });
+      setAppointments(data);
+    })();
+  }, [api, date, refreshCount]);
 
   return (
-    <Wrapper>
-      <TopBar
-        title={
-          <TranslatedText
-            stringId="scheduling.locationBookings.title"
-            fallback="Location bookings"
+    <PageContainer>
+      <TwoColumnDisplay>
+        <LeftContainer>
+          <TopBarBase
+            title={
+              <TranslatedText stringId="scheduling.appointmentCalendar.title" fallback="Calendar" />
+            }
           />
-        }
-      >
-        <Filters>
-          <Placeholder>Search</Placeholder>
-          <Placeholder>Area</Placeholder>
-          <Placeholder>Clinician</Placeholder>
-          <Placeholder>Type</Placeholder>
-        </Filters>
-      </TopBar>
-      <CalendarGrid
-        style={{
-          '--location-count': locations.length,
-          '--week-count': 6,
-        }}
-      >
-        {Array.from({ length: locations.length * 6 }, (_, i) => (
-          <CalendarCell key={i} />
-        ))}
-      </CalendarGrid>
-      {/* <Placeholder>LocationBookingsView</Placeholder> */}
-    </Wrapper>
+          <Section>
+            <SectionTitle variant="subtitle2">
+              <TranslatedText
+                stringId="scheduling.appointmentCalendar.subTitle"
+                fallback="View calendar by"
+              />
+              :
+            </SectionTitle>
+            <FilterSwitch>
+              {Object.entries(filters).map(([key, { label }]) => (
+                <Button
+                  key={key}
+                  color={key === activeFilter ? 'primary' : null}
+                  variant={key === activeFilter ? 'contained' : null}
+                  onClick={() => {
+                    setActiveFilter(key);
+                    setFilterValue('');
+                  }}
+                >
+                  {label}
+                </Button>
+              ))}
+            </FilterSwitch>
+          </Section>
+          <Section>
+            <SectionTitle variant="subtitle2">{filters[activeFilter].label}</SectionTitle>
+            {filters[activeFilter].component}
+          </Section>
+          <Section>
+            <SectionTitle variant="subtitle2">
+              <TranslatedText stringId="appointment.type.label" fallback="Appointment type" />
+            </SectionTitle>
+            <TranslatedMultiSelectField
+              onChange={e => {
+                if (!e.target.value) {
+                  setAppointmentType([]);
+                  return;
+                }
+                setAppointmentType(JSON.parse(e.target.value));
+              }}
+              value={appointmentType}
+              name="appointmentType"
+              enumValues={APPOINTMENT_TYPE_LABELS}
+            />
+          </Section>
+        </LeftContainer>
+        <RightContainer>
+          <TopBar>
+            <DateHeader>
+              <TodayButton
+                variant="text"
+                onClick={() => {
+                  setDate(new Date());
+                }}
+              >
+                <TranslatedText
+                  stringId="scheduling.appointmentCalendar.action.today"
+                  fallback="Today"
+                />
+              </TodayButton>
+
+              <NavigationButton
+                onClick={() => {
+                  setDate(add(date, { days: -1 }));
+                }}
+              >
+                <ArrowBackIcon />
+              </NavigationButton>
+              <DateDisplay>{format(date, 'EEEE dd MMMM yyyy')}</DateDisplay>
+              <NavigationButton
+                onClick={() => {
+                  setDate(add(date, { days: 1 }));
+                }}
+              >
+                <ArrowForwardIcon />
+              </NavigationButton>
+            </DateHeader>
+            <NewAppointmentButton onSuccess={updateCalendar} />
+          </TopBar>
+          <CalendarContainer>
+            {!hasPermission ? (
+              <ErrorMessage
+                title={
+                  <TranslatedText stringId="general.error.noPermission" fallback="No permission" />
+                }
+                errorMessage={
+                  <TranslatedText
+                    stringId="schedule.appointmentCalendar.error.noPermission"
+                    fallback="You do not have permission to view the appointments calendar. If you require access, please contact your administrator."
+                  />
+                }
+              />
+            ) : (
+              <DailySchedule
+                appointments={appointments}
+                activeFilter={activeFilter}
+                filterValue={filterValue}
+                appointmentType={appointmentType}
+                onAppointmentUpdated={updateCalendar}
+              />
+            )}
+          </CalendarContainer>
+        </RightContainer>
+      </TwoColumnDisplay>
+    </PageContainer>
   );
 };
